Updated local-cron to distributable package
- Yay
