+++ .gitignore
@@ -12,3 +12,4 @@
 *.DS_store
 *-lock.json
 *.lock
+*.tar.gz
