--- local-cron/Gemfile
@@ -1,4 +1,0 @@
-source 'https://rubygems.org'
-
-gem 'oauth'
-gem 'json'
