+++ local-cron/bin/CoreLocationCLI
@@ -1,0 +1,1 @@
+/usr/local/Caskroom/corelocationcli/3.0.1/CoreLocationCLI
