--- local-cron/config/slack.yml.example
@@ -1,28 +1,0 @@
-##############################################################################################
-#  Config file for the slack-status-updater
-#
-#  - Get a Slack oauth-token
-#    More info about where to get it: https://bitbucket.org/mantebridts/slackupdatestatus
-#
-#  - Add locations (address)
-#    - Choose an emoji (https://www.webpagefx.com/tools/emoji-cheat-sheet/)
-#    - Choose a custom message
-##############################################################################################
-
-oauth_key: "YOUR_TOKEN_HERE"
-location1:
-  address:
-    - Antwerp
-  slack:
-    message: "a custom message for location 1"
-    emoji: ":location1_emoji:"
-location2:
-  address:
-    - Ghent
-  slack:
-    message: "a custom message for location 2"
-    emoji: ":location2_emoji:"
-else:
-  slack:
-    message: "At a coffee shop"
-    emoji: "coffee:"
