+++ local-cron/lib/app/slack_location_status.rb
@@ -1,0 +1,4 @@
+$LOAD_PATH.unshift(File.expand_path(File.dirname(__FILE__)))
+
+require 'slack_location_status/update_status'
+require 'slack_location_status/status_from_location'
