--- local-cron/lib/slack_location_status.rb
@@ -1,5 +1,0 @@
-$LOAD_PATH.unshift(File.expand_path(File.dirname(__FILE__)))
-
-require 'slack_location_status/config'
-require 'slack_location_status/update_status'
-require 'slack_location_status/status_from_location'
