--- local-cron/lib/slack_location_status/status_from_location.rb
@@ -1,16 +1,0 @@
-module SlackLocationStatus
-  class StatusFromLocation
-    def self.location
-      @location ||= `/usr/local/bin/CoreLocationCLI -format "%address"`
-    end
-
-    def self.set_status
-      config = SlackLocationStatus::Config.config
-      puts "Location used:"
-      puts "--------------"
-      puts self.location
-      puts "--------------"      
-      SlackLocationStatus::UpdateStatus.post({address: self.location})
-    end
-  end
-end
